Add a coroutine called async_generator that will loop 10 times, each time asynchronously wait 1 second, then yield a random number between 0 and 10

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -11,7 +11,7 @@ import random
 import asyncio
 
 
-async def async_generator() -> AsyncGenerator[float, None]:
+async def async_generator() -> AsyncGenerator[int, None]:
     """
     coroutine called async_generator that takes no arguments
     """
